Refactor Validator
